Replaced angry transvestite with old guy named Albert Added various speeds of movement and transitioned between them Locked camera to player's head

+++ Unity/Assets/Standard Assets/Characters/FirstPersonCharacter/Scripts/RigidbodyFirstPersonController.cs
@@ -189,7 +189,7 @@ namespace UnityStandardAssets.Characters.FirstPerson
             }
             m_Jump = false;
 
-            animationController.SetFloat("moveSpeed", movementSettings.CurrentTargetSpeed);
+            animationController.SetInteger("moveSpeed", (int)(movementSettings.CurrentTargetSpeed));
         }
